XXX: Support to add custom headers for DoH upstream
Implement the '--upstream-http-header' new command argument to specify
the custom HTTP headers for DoH upstream.  This feature may be used to
attach the client tokens to achieve authentication at the upstream.
For example, the YamuCloud DoH server requires the 'X-DNS-Token' header
to authenticate the clients.

This argument may be specified multiple times to set multiple headers.